+++ packages/docs/src/pages/TooltipPage.tsx
@@ -5,7 +5,7 @@ import { Tooltip } from '@siakit/tooltip'
 export function TooltipPage() {
   return (
     <Flex flex align="center" justify="center">
-      <Tooltip content="this is tooltip">
+      <Tooltip content="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus in nisl mi.">
         <Button>button</Button>
       </Tooltip>
     </Flex>